+++ bundles/desktop/bundle.toml
@@ -125,8 +125,11 @@ to = "/system/scarlet/bin/ansi_banner"
 
 [[layers]]
 kind = "cargo"
-source = "../../user/bin"
+source = "../../user/video_player"
+package = "video_player"
 bin = "video_player"
+default-features = false
+features = ["h264-stateful-hw", "mp4-aac"]
 to = "/system/scarlet/bin/video_player"
 
 [[layers]]
