+++ user/video_player/Makefile.toml
@@ -1,0 +1,141 @@
+[config]
+default_to_workspace = false
+skip_core_tasks = true
+
+[env]
+OSABI_VALUE = "83"
+DIST_DIR = "../bin/dist/riscv64"
+DEBUG_DIR = "../bin/dist/riscv64.debug"
+TARGET_SPEC = "../targets/riscv64gc-unknown-scarlet-elf.json"
+
+[tasks.fmt]
+description = "Format video_player"
+command = "cargo"
+args = ["fmt"]
+
+[tasks.fmt-check]
+description = "Check video_player formatting"
+command = "cargo"
+args = ["fmt", "--", "--check"]
+
+[tasks.clippy]
+description = "Run clippy for video_player"
+command = "cargo"
+args = ["clippy", "--", "-D", "warnings"]
+
+[tasks.clippy-fix]
+description = "Run clippy with auto-fix for video_player"
+command = "cargo"
+args = ["clippy", "--fix", "--allow-dirty", "--allow-staged", "--", "-D", "warnings"]
+
+[tasks.build]
+description = "Build video_player for RISC-V in debug mode"
+dependencies = ["build-only-debug", "set-osabi-debug", "split-debug-symbols"]
+
+[tasks.build-release]
+description = "Build video_player for RISC-V in release mode"
+dependencies = ["build-only-release", "set-osabi-release"]
+
+[tasks.build-aarch64]
+description = "Build video_player for AArch64 in debug mode"
+dependencies = ["build-only-debug-aarch64", "set-osabi-debug-aarch64", "split-debug-symbols-aarch64"]
+
+[tasks.build-release-aarch64]
+description = "Build video_player for AArch64 in release mode"
+dependencies = ["build-only-release-aarch64", "set-osabi-release-aarch64"]
+
+[tasks.build-only-debug]
+description = "Build video_player for RISC-V without post-processing"
+script = [
+    'mkdir -p "${DIST_DIR}"',
+    'CC_riscv64gc_unknown_scarlet_elf="${PWD}/../../tools/riscv64-clang" cargo build --target "${TARGET_SPEC}" --out-dir "${DIST_DIR}"'
+]
+
+[tasks.build-only-release]
+description = "Build video_player for RISC-V release without post-processing"
+script = [
+    'mkdir -p "${DIST_DIR}"',
+    'CC_riscv64gc_unknown_scarlet_elf="${PWD}/../../tools/riscv64-clang" cargo build --release --target "${TARGET_SPEC}" --out-dir "${DIST_DIR}"'
+]
+
+[tasks.build-only-debug-aarch64]
+description = "Build video_player for AArch64 without post-processing"
+env = { "DIST_DIR" = "../bin/dist/aarch64", "DEBUG_DIR" = "../bin/dist/aarch64.debug", "TARGET_SPEC" = "../targets/aarch64-unknown-scarlet-elf.json" }
+script = [
+    'mkdir -p "${DIST_DIR}"',
+    'cargo build --target "${TARGET_SPEC}" --out-dir "${DIST_DIR}"'
+]
+
+[tasks.build-only-release-aarch64]
+description = "Build video_player for AArch64 release without post-processing"
+env = { "DIST_DIR" = "../bin/dist/aarch64", "DEBUG_DIR" = "../bin/dist/aarch64.debug", "TARGET_SPEC" = "../targets/aarch64-unknown-scarlet-elf.json" }
+script = [
+    'mkdir -p "${DIST_DIR}"',
+    'cargo build --release --target "${TARGET_SPEC}" --out-dir "${DIST_DIR}"'
+]
+
+[tasks.set-osabi-debug]
+description = "Set OSABI for video_player RISC-V debug"
+dependencies = ["build-only-debug"]
+script = 'python3 ../bin/set_osabi.py "${DIST_DIR}/video_player" ${OSABI_VALUE}'
+
+[tasks.set-osabi-release]
+description = "Set OSABI for video_player RISC-V release"
+dependencies = ["build-only-release"]
+script = 'python3 ../bin/set_osabi.py "${DIST_DIR}/video_player" ${OSABI_VALUE}'
+
+[tasks.set-osabi-debug-aarch64]
+description = "Set OSABI for video_player AArch64 debug"
+env = { "DIST_DIR" = "../bin/dist/aarch64" }
+dependencies = ["build-only-debug-aarch64"]
+script = 'python3 ../bin/set_osabi.py "${DIST_DIR}/video_player" ${OSABI_VALUE}'
+
+[tasks.set-osabi-release-aarch64]
+description = "Set OSABI for video_player AArch64 release"
+env = { "DIST_DIR" = "../bin/dist/aarch64" }
+dependencies = ["build-only-release-aarch64"]
+script = 'python3 ../bin/set_osabi.py "${DIST_DIR}/video_player" ${OSABI_VALUE}'
+
+[tasks.split-debug-symbols]
+description = "Split debug symbols for video_player RISC-V debug"
+env = { "DEBUG_DIR" = "../bin/dist/riscv64.debug", "DIST_DIR" = "../bin/dist/riscv64" }
+script = '''
+mkdir -p "${DEBUG_DIR}"
+if command -v llvm-objcopy >/dev/null 2>&1; then
+    OBJCOPY=llvm-objcopy
+elif command -v riscv64-unknown-elf-objcopy >/dev/null 2>&1; then
+    OBJCOPY=riscv64-unknown-elf-objcopy
+elif command -v riscv64-linux-gnu-objcopy >/dev/null 2>&1; then
+    OBJCOPY=riscv64-linux-gnu-objcopy
+else
+    echo "Error: objcopy not found for RISC-V"
+    exit 1
+fi
+binary_path="${DIST_DIR}/video_player"
+debug_file="${DEBUG_DIR}/video_player.debug"
+cp "${binary_path}" "${binary_path}.tmp" && mv "${binary_path}.tmp" "${binary_path}"
+"$OBJCOPY" --only-keep-debug "${binary_path}" "${debug_file}"
+"$OBJCOPY" --strip-debug --strip-unneeded "${binary_path}"
+"$OBJCOPY" --add-gnu-debuglink="${debug_file}" "${binary_path}"
+'''
+
+[tasks.split-debug-symbols-aarch64]
+description = "Split debug symbols for video_player AArch64 debug"
+env = { "DEBUG_DIR" = "../bin/dist/aarch64.debug", "DIST_DIR" = "../bin/dist/aarch64" }
+script = '''
+mkdir -p "${DEBUG_DIR}"
+if command -v llvm-objcopy >/dev/null 2>&1; then
+    OBJCOPY=llvm-objcopy
+elif command -v aarch64-linux-gnu-objcopy >/dev/null 2>&1; then
+    OBJCOPY=aarch64-linux-gnu-objcopy
+else
+    echo "Warning: objcopy not found for AArch64, skipping debug symbol split"
+    exit 0
+fi
+binary_path="${DIST_DIR}/video_player"
+debug_file="${DEBUG_DIR}/video_player.debug"
+cp "${binary_path}" "${binary_path}.tmp" && mv "${binary_path}.tmp" "${binary_path}"
+"$OBJCOPY" --only-keep-debug "${binary_path}" "${debug_file}"
+"$OBJCOPY" --strip-debug --strip-unneeded "${binary_path}"
+"$OBJCOPY" --add-gnu-debuglink="${debug_file}" "${binary_path}"
+'''
